Finish formatting methods
