+++ src/containers/Outcomes/OutcomesDetail.tsx
@@ -1,10 +1,8 @@
 import React from 'react'
-import { Message, Header, Form, Segment } from 'semantic-ui-react'
+import { Header, Form, Segment } from 'semantic-ui-react'
 import { Loading } from '../../components'
 import usePageTitle from '../../utils/hooks/usePageTitle'
 import { useRouter } from '../../utils/hooks/useRouter'
-import { SummaryViewWrapper } from '../../formElementPlugins'
-import strings from '../../utils/constants'
 import { TemplateElementCategory } from '../../utils/generated/graphql'
 import {
   ApplicationDetails,
@@ -15,7 +13,7 @@ import {
 import config from '../../config'
 import { defaultEvaluatedElement } from '../../utils/hooks/useLoadApplication'
 import { useOutcomesDetail } from '../../utils/hooks/useOutcomes'
-import { FormattedCell, formatCellText } from './FormattedCell'
+import { constructElement } from './helpers'
 import { DisplayDefinition } from './types'
 
 const OutcomeDetails: React.FC = () => {
@@ -47,55 +45,6 @@ const OutcomeDetails: React.FC = () => {
       </Form>
     </>
   )
-}
-
-const constructElement = (value: any, displayDefinition: DisplayDefinition, id: number) => {
-  const { title } = displayDefinition
-  const { elementTypePluginCode, elementParameters, response } = getElementDetails(
-    value,
-    displayDefinition
-  )
-  const element = {
-    id,
-    code: `Detail${id}`,
-    pluginCode: elementTypePluginCode as string,
-    category: TemplateElementCategory.Information,
-    title,
-    parameters: elementParameters,
-    validationExpression: true,
-    validationMessage: '',
-    elementIndex: 0,
-    page: 0,
-    sectionIndex: 0,
-    helpText: null,
-    sectionCode: '0',
-    ...defaultEvaluatedElement,
-    isRequired: false,
-  }
-  return (
-    <SummaryViewWrapper
-      element={element}
-      applicationData={{ config } as ApplicationDetails}
-      /* TODO this is a hacky way of passing through server URL, I think it needs to be decoupled from applicationData. Config is needed for log_url in organisation to be displayed with imageDisplays plugin */
-      response={response}
-      allResponses={{}}
-    />
-  )
-}
-
-const getElementDetails = (value: any, displayDefinition: DisplayDefinition) => {
-  // If it's not already a plugin element, structure it as a shortText
-  // element for display purposes
-  const { formatting } = displayDefinition
-  const isAlreadyElement = !!formatting?.elementTypePluginCode
-  const elementTypePluginCode = isAlreadyElement ? formatting?.elementTypePluginCode : 'shortText'
-  const elementParameters = isAlreadyElement
-    ? formatting?.elementParameters
-    : {
-        label: displayDefinition.title,
-      }
-  const response = isAlreadyElement ? value : { text: formatCellText(value, displayDefinition) }
-  return { elementTypePluginCode, elementParameters, response }
 }
 
 export default OutcomeDetails
