--- src/utils/graphql/queries/getOutcomeDisplays.query.ts
@@ -1,34 +1,0 @@
-import { gql } from '@apollo/client'
-
-export default gql`
-  query getOutcomeDisplays {
-    outcomeDisplays {
-      nodes {
-        code
-        detailColumnName
-        id
-        pluralTableName
-        tableName
-        title
-        outcomeDisplayTables {
-          nodes {
-            columnName
-            id
-            isTextColumn
-            title
-          }
-        }
-        outcomeDisplayDetails {
-          nodes {
-            columnName
-            elementTypePluginCode
-            isTextColumn
-            id
-            title
-            parameters
-          }
-        }
-      }
-    }
-  }
-`
